refactor: introduce page
